pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "yadirmenu"
version = "0.1"
authors = [
  { name="Jevgenij Kazanov", email="evgeny.kazanov@gmail.com" },
]
dependencies = [
  "pythondialog",
  "python-dotenv"
]
description = "Make a menu tree from directory tree."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/ekazanov/yadirmenu"
